// app/requests/_personal/getManager/getManagerSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetManagerSuccess {
  data: {
    id: number;
    phone: number;
    email: string;
    logo?: null | string;
    roles: {
      id: number;
      name:
        | "admin"
        | "client"
        | "manager"
        | "recruiter"
        | "specialist"
        | "supervisor";
      [k: string]: unknown;
    }[];
    [k: string]: unknown;
  }[];
  [k: string]: unknown;
}